fix linters, move injectors to pkg
